requirements.txt
fastapi==0.104.1
uvicorn[standard]>=0.27.0
websockets>=15.0
python-dotenv==1.0.0
sarvamai
pydantic>=2.4.0
python-multipart==0.0.6

